+++ servers/Azure.Mcp.Server/CHANGELOG.md
@@ -7,7 +7,7 @@ The Azure MCP Server updates automatically by default whenever a new release com
 ### Features Added
 
 - Added the following Azure Managed Lustre commands: [[#110](https://github.com/microsoft/mcp/issues/110)]
-  - `azmcp_azuremanagedlustre_filesystem_check-subnet-size`: Check if the subnet can host the target Azure Managed Lustre SKU and size.
+  - `azmcp_azuremanagedlustre_filesystem_subnet-size-validate`: Check if the subnet can host the target Azure Managed Lustre SKU and size.
 
 ### Breaking Changes
 
@@ -61,7 +61,7 @@ The Azure MCP Server updates automatically by default whenever a new release com
 - Added support for getting details of an Azure Function App via the command `azmcp_functionapp_get`. [[#970](https://github.com/Azure/azure-mcp/pull/970)]
 - Added the following Azure Managed Lustre commands: [[#1003](https://github.com/Azure/azure-mcp/issues/1003)]
   - `azmcp_azuremanagedlustre_filesystem_list`: List available Azure Managed Lustre filesystems.
-  - `azmcp_azuremanagedlustre_filesystem_required-subnet-size`: Returns the number of IP addresses required for a specific SKU and size of Azure Managed Lustre filesystem.
+  - `azmcp_azuremanagedlustre_filesystem_subnet-size-required`: Returns the number of IP addresses required for a specific SKU and size of Azure Managed Lustre filesystem.
 - Added support for designing Azure Cloud Architecture through guided questions via the command `azmcp_cloudarchitect_design`. [[#890](https://github.com/Azure/azure-mcp/pull/890)]
 - Added support for the following Azure MySQL operations: [[#855](https://github.com/Azure/azure-mcp/issues/855)]
   - `azmcp_mysql_database_list` - List all databases in a MySQL server.